Added Query1 uniform job
